requirements.txt
Django==3.1.7
requests
django-colorfield